requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
docstring_parser==0.16
exceptiongroup==1.2.1
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.1
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
selenium==4.21.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vertexai==1.49.0
wsproto==1.2.0
yarl==1.9.4
